# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AIOManager API",
    "description": "Hybrid sync server for AIOManager. Served same-origin under /api/*.",
    "version": "2.0.0"
  },
...
  "paths": {
...
    "/api/commands": {
      "post": {
        "summary": "Enqueue a bulk account command (protect-all, unprotect-all)",
        "tags": [
          "commands"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id",
                  "accountIds"
                ],
                "properties": {
                  "id": {
                    "type": "string",
                    "enum": [
                      "protect-all",
                      "unprotect-all"
                    ],
                    "description": "Command id (matches the client bulk-action registry)"
                  },
                  "accountIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1
                    },
                    "minItems": 1,
                    "maxItems": 200,
                    "description": "Account ids the command should run against"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "jobId",
                    "status"
                  ],
                  "properties": {
                    "jobId": {
                      "type": "string",
                      "description": "Job id for polling GET /api/commands/:jobId"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "queued"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/commands/{jobId}": {
      "get": {
        "summary": "Poll a command job status and per-account results",
        "tags": [
          "commands"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "jobId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "jobId",
                    "command",
                    "status"
                  ],
                  "properties": {
                    "jobId": {
                      "type": "string"
                    },
                    "command": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "queued",
                        "running",
                        "completed",
                        "failed"
                      ]
                    },
                    "results": {
                      "type": [
                        "null",
                        "array"
                      ],
                      "items": {
                        "type": "object",
                        "required": [
                          "accountId",
                          "ok"
                        ],
                        "properties": {
                          "accountId": {
                            "type": "string"
                          },
                          "ok": {
                            "type": "boolean"
                          },
                          "error": {
                            "type": "string",
                            "description": "Present when ok is false"
                          },
                          "changed": {
                            "type": "number",
                            "description": "Addons whose protection flag changed"
                          },
                          "total": {
                            "type": "number",
                            "description": "Addons in the account collection"
                          }
                        }
                      }
                    },
                    "error": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "number",
                      "description": "Unix epoch (ms)"
                    },
                    "updatedAt": {
                      "type": "number",
                      "description": "Unix epoch (ms)"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
}